requirements.txt
redis>=5.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
